Refactor monetary fields to use BigDecimal instead of Double across DTOs, entities, services, and tests for improved precision in financial calculations.

+++ src/main/java/com/vaPaTi/vaPaTi/dtos/CampaignResponseDTO.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -15,8 +16,8 @@ public class CampaignResponseDTO {
     private Long id;
     private String name;
     private String description;
-    private Double amountGoal;
-    private Double amountRaised;
+    private BigDecimal amountGoal;
+    private BigDecimal amountRaised;
     private Long userId;
     private List<CategoryDTO> categories;
     private CampaignStatus status;
